+++ README.md
@@ -37,8 +37,11 @@
     - 提供单独重启某一策略实例功能，可在线更新策略源码后，重启某一策略实例，不影响其他运行实例。
     - 支持单策略多合约行情订阅，支持指数合约行情订阅
     - 提供组合回测引擎，能够直接加载cta_strategy_pro_setting.json文件进行组合回测。
+    - 拆分组合回测引擎和回测引擎，组合回测引擎支持bar/tick级别的组合回测
     - 增加定时器，推动策略on_timer
     - 增加定时推送策略持仓event
+    - 增加CtaPro模板，支持精细化策略持久模板，
+    - 增加CtaPro期货模板，支持FAK委托，自动换月等
     
 8、增强主引擎，包括：
