+++ tests/renko/test_rebuild_stock.py
