+++ vnpy/app/cta_backtester/engine.py
@@ -4,7 +4,6 @@ import traceback
 from datetime import datetime
 from threading import Thread
 from pathlib import Path
-from inspect import getfile
 
 from vnpy.event import Event, EventEngine
 from vnpy.trader.engine import BaseEngine, MainEngine
