+++ vnpy/data/renko/config.py
@@ -3,5 +3,5 @@
 
 HEIGHT_LIST = [3, 5, 10, 'K3', 'K5', 'K10']
 
-FUTURE_RENKO_DB_NAME = 'FutureRenko_Db'
-STOCK_RENKO_DB_NAME = 'StockRenko_Db'
+FUTURE_RENKO_DB_NAME = 'FutureRenko'
+STOCK_RENKO_DB_NAME = 'StockRenko'
