+++ win_clean_celery_jobs.bat
@@ -1,0 +1,1 @@
+celery -A vnpy.task.celery_app purge
