+++ win_start_celery_worker.bat
@@ -1,0 +1,1 @@
+celery worker -c 2 -A vnpy.task.celery_app  -P eventlet -l debug -f celery.log
